Create CODEOWNERS
We now require all repositories within our organization to have CODEOWNERS defined.
I see this repository has a OWNERS file defined, but could not find any documentation on Github that suggests that this is supported or what the difference is compared to CODEOWNERS. Could it be obsolete? Either way the latter would server the same purpose.